Pass struct generation logic to DataIterator
