+++ modules/ballerina-core/src/main/java/org/ballerinalang/model/ColumnDefinition.java
@@ -1,0 +1,44 @@
+/*
+ *  Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+package org.ballerinalang.model;
+
+import org.ballerinalang.model.types.TypeKind;
+
+/**
+ * This  wraps the column name and type of a selected result set from a data source.
+ *
+ * @since 0.95.0
+ */
+public abstract class ColumnDefinition {
+
+    protected String name;
+    protected TypeKind mappedType;
+
+    public ColumnDefinition(String name, TypeKind mappedType) {
+        this.name = name;
+        this.mappedType = mappedType;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public TypeKind getType() {
+        return mappedType;
+    }
+}
