+++ modules/ballerina-core/src/main/java/org/ballerinalang/model/DataTableJSONDataSource.java
@@ -24,7 +24,6 @@ import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.ballerinalang.model.values.BDataTable;
-import org.ballerinalang.model.values.BDataTable.ColumnDefinition;
 import org.ballerinalang.model.values.BJSON.JSONDataSource;
 
 import java.io.IOException;
