+++ modules/ballerina-test/src/test/resources/test-src/types/datatable/datatable-type.bal
@@ -413,3 +413,25 @@ function testCloseConnectionPool () (int count) {
     testDB.close();
     return;
 }
+
+function testMutltipleRows () (int i1, int i2) {
+    sql:ClientConnector testDB = create sql:ClientConnector(sql:HSQLDB_FILE, "./target/tempdb/",
+                                                            0, "TEST_DATA_TABLE_DB", "SA", "", {maximumPoolSize:1});
+
+    sql:Parameter[] parameters = [];
+    datatable dt = testDB.select("SELECT int_type from DataTableRep", parameters);
+    ResultPrimitiveInt rs1;
+    ResultPrimitiveInt rs2;
+    int i = 0;
+    while (dt.hasNext()) {
+        any dataStruct = dt.getNext();
+        if (i == 0) {
+            rs1, _ = (ResultPrimitiveInt)dataStruct;
+        } else {
+            rs2, _ = (ResultPrimitiveInt)dataStruct;
+        }
+        i = i +1;
+    }
+    testDB.close();
+    return rs1.INT_TYPE, rs2.INT_TYPE;
+}
